+++ frontend/src/App.vue
@@ -2,9 +2,9 @@
 import { provide, ref } from "vue";
 import { RouterView, useRoute, useRouter } from "vue-router";
 import { Tooltip } from "bootstrap";
-import AppFooter from "./components/Layout/AppFooter.vue";
-import AppNavbar from "./components/Layout/AppNavbar.vue";
-import ToastContainer from "./components/Layout/ToastContainer.vue";
+import AppFooter from "./components/_layout/AppFooter.vue";
+import AppNavbar from "./components/_layout/AppNavbar.vue";
+import ToastContainer from "./components/_layout/ToastContainer.vue";
 import type { ApiError } from "./types/ApiError";
 import type { ToastOptions } from "./types/ToastOptions";
 import type { ToastUtils } from "./types/ToastUtils";
