+++ frontend/src/components/_layout/AppFooter.vue
