+++ frontend/src/components/_layout/AppNavbar.vue
