+++ frontend/src/components/_users/ContactInformation.vue
@@ -5,11 +5,11 @@ import { useI18n } from "vue-i18n";
 import AddressLineInput from "./AddressLineInput.vue";
 import AddressLocalityInput from "./AddressLocalityInput.vue";
 import CountrySelect from "./CountrySelect.vue";
-import EmailAddressInput from "@/components/Users/EmailAddressInput.vue";
-import PhoneExtensionInput from "@/components/Users/PhoneExtensionInput.vue";
-import PhoneNumberInput from "@/components/Users/PhoneNumberInput.vue";
-import PostalCodeInput from "@/components/Users/PostalCodeInput.vue";
-import RegionSelect from "@/components/Users/RegionSelect.vue";
+import EmailAddressInput from "@/components/_users/EmailAddressInput.vue";
+import PhoneExtensionInput from "@/components/_users/PhoneExtensionInput.vue";
+import PhoneNumberInput from "@/components/_users/PhoneNumberInput.vue";
+import PostalCodeInput from "@/components/_users/PostalCodeInput.vue";
+import RegionSelect from "@/components/_users/RegionSelect.vue";
 import type { AddressInput } from "@/types/AddressInput";
 import type { CountrySettings } from "@/types/CountrySettings";
 import type { EmailInput } from "@/types/EmailInput";
