+++ frontend/src/components/_users/EmailAddressInput.vue
