+++ frontend/src/components/_users/PictureInput.vue
