+++ frontend/src/views/_realms/RealmEdit.vue
@@ -3,10 +3,10 @@ import { computed, inject, onMounted, ref } from "vue";
 import { useForm } from "vee-validate";
 import { useI18n } from "vue-i18n";
 import { useRoute, useRouter } from "vue-router";
-import ClaimMappingList from "@/components/Realms/ClaimMappingList.vue";
-import JwtSecretField from "@/components/Realms/JwtSecretField.vue";
-import PasswordSettingsEdit from "@/components/Realms/PasswordSettingsEdit.vue";
-import UsernameSettingsEdit from "@/components/Realms/UsernameSettingsEdit.vue";
+import ClaimMappingList from "@/components/_realms/ClaimMappingList.vue";
+import JwtSecretField from "@/components/_realms/JwtSecretField.vue";
+import PasswordSettingsEdit from "@/components/_realms/PasswordSettingsEdit.vue";
+import UsernameSettingsEdit from "@/components/_realms/UsernameSettingsEdit.vue";
 import type { ClaimMapping } from "@/types/ClaimMapping";
 import type { CustomAttribute } from "@/types/CustomAttribute";
 import type { PasswordSettings } from "@/types/PasswordSettings";
