+++ frontend/src/views/_users/SignIn.vue
@@ -3,8 +3,8 @@ import { inject, ref } from "vue";
 import { useForm } from "vee-validate";
 import { useI18n } from "vue-i18n";
 import { useRoute, useRouter } from "vue-router";
-import PasswordInput from "@/components/Users/PasswordInput.vue";
-import UsernameInput from "@/components/Users/UsernameInput.vue";
+import PasswordInput from "@/components/_users/PasswordInput.vue";
+import UsernameInput from "@/components/_users/UsernameInput.vue";
 import type { ApiError } from "@/types/ApiError";
 import type { ErrorDetail } from "@/types/ErrorDetail";
 import { handleErrorKey } from "@/inject/App";
