+++ frontend/src/views/_users/SignOut.vue
